+++ tests/service/e2e/cucumber/steps/envelope_steps.go
@@ -504,6 +504,45 @@ func (sc *ScenarioContext) iSetNonceRecords(table *gherkin.PickleStepArgument_Pi
 	return nil
 }
 
+func (sc *ScenarioContext) iSetNonceLastAttributedRecords(table *gherkin.PickleStepArgument_PickleTable) error {
+	aliasTable := utils.ExtractColumns(table, []string{"Account", "ChainID", "Nonce"})
+	if aliasTable == nil {
+		return errors.DataError("Missing mandatory columns")
+	}
+
+	for _, row := range aliasTable.Rows[1:] {
+		acc := row.Cells[0].Value
+		chainID := row.Cells[1].Value
+		nonce, err := strconv.ParseInt(row.Cells[2].Value, 10, 64)
+		if err != nil {
+			return err
+		}
+
+		nonceKey := fmt.Sprintf("%v@%v", acc, chainID)
+		if nonce >= 0 {
+			if err := sc.nonceManager.SetLastAttributed(nonceKey, uint64(nonce)); err != nil {
+				return err
+			}
+
+			log.WithFields(log.Fields{
+				"SetLastAttributed": nonce,
+				"nonce_key":         nonceKey,
+			}).Debug("scenario: set nonce last attribute records")
+		} else {
+			if err := sc.nonceManager.DeleteLastAttributed(nonceKey); err != nil {
+				return err
+			}
+
+			log.WithFields(log.Fields{
+				"SetLastAttributed": nonce,
+				"nonce_key":         nonceKey,
+			}).Debug("scenario: delete nonce last attribute records")
+		}
+	}
+
+	return nil
+}
+
 func (sc *ScenarioContext) iTrackTheFollowingEnvelopes(table *gherkin.PickleStepArgument_PickleTable) error {
 	if len(table.Rows[0].Cells) != 1 {
 		return errors.DataError("invalid table")
@@ -616,6 +655,7 @@ func initEnvelopeSteps(s *godog.ScenarioContext, sc *ScenarioContext) {
 	s.Step(`^I have the following account`, sc.preProcessTableStep(sc.iHaveTheFollowingAccount))
 	s.Step(`^I register the following alias$`, sc.preProcessTableStep(sc.iRegisterTheFollowingAliasAs))
 	s.Step(`^Set nonce manager records$`, sc.preProcessTableStep(sc.iSetNonceRecords))
+	s.Step(`^Set nonce last attributed records$`, sc.preProcessTableStep(sc.iSetNonceLastAttributedRecords))
 	s.Step(`^I have created the following accounts$`, sc.preProcessTableStep(sc.iHaveCreatedTheFollowingAccounts))
 	s.Step(`^I track the following envelopes$`, sc.preProcessTableStep(sc.iTrackTheFollowingEnvelopes))
 	s.Step(`^I send envelopes to topic "([^"]*)"$`, sc.iSendEnvelopesToTopic)